requirements.txt
Package              Version
-------------------- -----------
annotated-types      0.5.0
anyio                3.7.1
certifi              2023.7.22
charset-normalizer   3.2.0
click                8.1.6
colorama             0.4.6
contourpy            1.1.0
cycler               0.11.0
dnspython            2.4.0
email-validator      2.0.0.post2
fastapi              0.100.0
filelock             3.12.2
fonttools            4.41.1
h11                  0.14.0
httpcore             0.17.3
httptools            0.6.0
httpx                0.24.1
idna                 3.4
itsdangerous         2.1.2
Jinja2               3.1.2
kiwisolver           1.4.4
MarkupSafe           2.1.3
matplotlib           3.7.2
mpmath               1.3.0
networkx             3.1
numpy                1.25.1
opencv-python        4.8.0.74
orjson               3.9.2
packaging            23.1
pandas               2.0.3
Pillow               10.0.0
pip                  23.2.1
psutil               5.9.5
py-cpuinfo           9.0.0
pydantic             2.0.3
pydantic_core        2.3.0
pydantic-extra-types 2.0.0
pydantic-settings    2.0.2
pyparsing            3.0.9
python-dateutil      2.8.2
python-dotenv        1.0.0
python-multipart     0.0.6
pytz                 2023.3
PyYAML               6.0.1
requests             2.31.0
scipy                1.11.1
seaborn              0.12.2
setuptools           60.2.0
six                  1.16.0
sniffio              1.3.0
starlette            0.27.0
sympy                1.12
torch                2.0.1
torchvision          0.15.2
tqdm                 4.65.0
typing_extensions    4.7.1
tzdata               2023.3
ujson                5.8.0
ultralytics          8.0.143
urllib3              2.0.4
uvicorn              0.23.1
watchfiles           0.19.0
websockets           11.0.3
wheel                0.37.1
